Move CannotFindMatch code to modules
